Squashfs: fix corruption checks in squashfs_readdir()
The dir_count and size fields when read from disk are sanity
checked for correctness.  However, the sanity checks only check the
values are not greater than expected.  As dir_count and size were
incorrectly defined as signed ints, this can lead to corrupted values
appearing as negative which are not trapped.

Change-Id: If44b279c1d012b2a7b1c659a6fba84f476247d96
Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>